refactor: add 'Pgn' to icon names that start with a number
