+++ icons/jsx/index.jsx
@@ -1,60 +1,60 @@
-export { default as TenK } from './10K';
-export { default as TenMp } from './10Mp';
-export { default as ElevenMp } from './11Mp';
-export { default as OneTwoThree } from './123';
-export { default as TwelveMp } from './12Mp';
-export { default as ThirteenMp } from './13Mp';
-export { default as FourteenMp } from './14Mp';
-export { default as FifteenMp } from './15Mp';
-export { default as SixteenMp } from './16Mp';
-export { default as SeventeenMp } from './17Mp';
-export { default as EighteenUpRating } from './18UpRating';
-export { default as EighteenMp } from './18Mp';
-export { default as NineteenMp } from './19Mp';
-export { default as OneK } from './1K';
-export { default as OneKPlus } from './1KPlus';
-export { default as OneXMobiledata } from './1XMobiledata';
-export { default as TwentyMp } from './20Mp';
-export { default as TwentyOneMp } from './21Mp';
-export { default as TwentyTwoMp } from './22Mp';
-export { default as TwentyThreeMp } from './23Mp';
-export { default as TwentyFourMp } from './24Mp';
-export { default as TwoK } from './2K';
-export { default as TwoKPlus } from './2KPlus';
-export { default as TwoMp } from './2Mp';
-export { default as ThirtyFps } from './30Fps';
-export { default as ThirtyFpsSelect } from './30FpsSelect';
-export { default as ThreeSixty } from './360';
-export { default as ThreeDRotation } from './3DRotation';
-export { default as ThreeGMobiledata } from './3GMobiledata';
-export { default as ThreeK } from './3K';
-export { default as ThreeKPlus } from './3KPlus';
-export { default as ThreeMp } from './3Mp';
-export { default as ThreeP } from './3P';
-export { default as FourGMobiledata } from './4GMobiledata';
-export { default as FourGPlusMobiledata } from './4GPlusMobiledata';
-export { default as FourK } from './4K';
-export { default as FourKPlus } from './4KPlus';
-export { default as FourMp } from './4Mp';
-export { default as FiveG } from './5G';
-export { default as FiveK } from './5K';
-export { default as FiveKPlus } from './5KPlus';
-export { default as FiveMp } from './5Mp';
-export { default as SixtyFps } from './60Fps';
-export { default as SixtyFpsSelect } from './60FpsSelect';
-export { default as SixFtApart } from './6FtApart';
-export { default as SixK } from './6K';
-export { default as SixKPlus } from './6KPlus';
-export { default as SixMp } from './6Mp';
-export { default as SevenK } from './7K';
-export { default as SevenKPlus } from './7KPlus';
-export { default as SevenMp } from './7Mp';
-export { default as EightK } from './8K';
-export { default as EightKPlus } from './8KPlus';
-export { default as EightMp } from './8Mp';
-export { default as NineK } from './9K';
-export { default as NineKPlus } from './9KPlus';
-export { default as NineMp } from './9Mp';
+export { default as Pgn10K } from './10K';
+export { default as Pgn10Mp } from './10Mp';
+export { default as Pgn11Mp } from './11Mp';
+export { default as Pgn123 } from './123';
+export { default as Pgn12Mp } from './12Mp';
+export { default as Pgn13Mp } from './13Mp';
+export { default as Pgn14Mp } from './14Mp';
+export { default as Pgn15Mp } from './15Mp';
+export { default as Pgn16Mp } from './16Mp';
+export { default as Pgn17Mp } from './17Mp';
+export { default as Pgn18UpRating } from './18UpRating';
+export { default as Pgn18Mp } from './18Mp';
+export { default as Pgn19Mp } from './19Mp';
+export { default as Pgn1K } from './1K';
+export { default as Pgn1KPlus } from './1KPlus';
+export { default as Pgn1XMobiledata } from './1XMobiledata';
+export { default as Pgn20Mp } from './20Mp';
+export { default as Pgn21Mp } from './21Mp';
+export { default as Pgn22Mp } from './22Mp';
+export { default as Pgn23Mp } from './23Mp';
+export { default as Pgn24Mp } from './24Mp';
+export { default as Pgn2K } from './2K';
+export { default as Pgn2KPlus } from './2KPlus';
+export { default as Pgn2Mp } from './2Mp';
+export { default as Pgn30Fps } from './30Fps';
+export { default as Pgn30FpsSelect } from './30FpsSelect';
+export { default as Pgn360 } from './360';
+export { default as Pgn3DRotation } from './3DRotation';
+export { default as Pgn3GMobiledata } from './3GMobiledata';
+export { default as Pgn3K } from './3K';
+export { default as Pgn3KPlus } from './3KPlus';
+export { default as Pgn3Mp } from './3Mp';
+export { default as Pgn3P } from './3P';
+export { default as Pgn4GMobiledata } from './4GMobiledata';
+export { default as Pgn4GPlusMobiledata } from './4GPlusMobiledata';
+export { default as Pgn4K } from './4K';
+export { default as Pgn4KPlus } from './4KPlus';
+export { default as Pgn4Mp } from './4Mp';
+export { default as Pgn5G } from './5G';
+export { default as Pgn5K } from './5K';
+export { default as Pgn5KPlus } from './5KPlus';
+export { default as Pgn5Mp } from './5Mp';
+export { default as Pgn60Fps } from './60Fps';
+export { default as Pgn60FpsSelect } from './60FpsSelect';
+export { default as Pgn6FtApart } from './6FtApart';
+export { default as Pgn6K } from './6K';
+export { default as Pgn6KPlus } from './6KPlus';
+export { default as Pgn6Mp } from './6Mp';
+export { default as Pgn7K } from './7K';
+export { default as Pgn7KPlus } from './7KPlus';
+export { default as Pgn7Mp } from './7Mp';
+export { default as Pgn8K } from './8K';
+export { default as Pgn8KPlus } from './8KPlus';
+export { default as Pgn8Mp } from './8Mp';
+export { default as Pgn9K } from './9K';
+export { default as Pgn9KPlus } from './9KPlus';
+export { default as Pgn9Mp } from './9Mp';
 export { default as Abc } from './Abc';
 export { default as AcUnit } from './AcUnit';
 export { default as AccessAlarm } from './AccessAlarm';
